pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paw-agent"
version = "0.1.0"
description = "Paw-Agent: local-first CLI coding agent for small models on llama.cpp"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
paw-agent = "paw_agent.cli:main"
paw = "paw_agent.cli:main"

[tool.setuptools]
packages = ["paw_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider -p no:tmpdir"
